Moved animation classes into separate package
